feat(bidi): add arg checks
